Make a copy of slice before sorting and return sorted slice
Fix test coverage issue
