+++ 02_bubble/mohankrishna/main_test.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"bytes"
-	"reflect"
+	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -27,24 +27,24 @@ var tests = []struct {
 	in  []int
 	out []int
 }{
+	{[]int{3, 2, 1, 5}, []int{1, 2, 3, 5}},
 	{[]int{3, 2, 1, 5, -1, -90, -123, 4387, 29384}, []int{-123, -90, -1, 1, 2, 3, 5, 4387, 29384}},
 	{[]int{8, 5, 3, 1, 9, 6, 0, 7, 4, 2, 5}, []int{0, 1, 2, 3, 4, 5, 5, 6, 7, 8, 9}},
 	{[]int{10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0}, []int{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10}}}
 
 func TestBubbleSortWithValues(t *testing.T) {
+	r := require.New(t)
 	for _, tt := range tests {
 		out := bubble(tt.in)
-		if !reflect.DeepEqual(out, tt.out) {
-			t.Errorf("got %v, want %v", out, tt.out)
-		}
+		r.ElementsMatch(tt.out, out)
 	}
 }
 
 func TestInsertionSortWithValues(t *testing.T) {
+	r := require.New(t)
 	for _, tt := range tests {
+		fmt.Println("insertion testing", tt.in, tt.out)
 		out := insertionSort(tt.in)
-		if !reflect.DeepEqual(out, tt.out) {
-			t.Errorf("got %v, want %v", out, tt.out)
-		}
+		r.ElementsMatch(tt.out, out)
 	}
 }
